refactor(ui): colocate admin-panel, organizations, router-settings, and old-usage page views
Continues the colocation follow-up. These four segments rendered their page
view out of the shared src/components dump; this moves each self-contained
view (and its colocated test, where one existed) into the owning route's
_components/ folder:

- src/components/AdminPanel.tsx        -> admin-panel/_components/
- src/components/organizations.tsx     -> organizations/_components/
- src/components/general_settings.tsx  -> router-settings/_components/
- src/components/usage.tsx             -> old-usage/_components/

Each view was verified to have no importers other than its own page (the only
other references were dead defensive mocks in the legacy CreateKeyPage test,
repointed here). Relative imports inside each moved file are rewritten to
absolute @/components/* since the genuinely shared dependencies (networking,
selectors, settings trees) stay in @/components. The two moved tests
(AdminPanel, organizations) have their vi.mock paths rewritten to match.

Grandfathered lint suppressions for the four files are re-keyed to their new
paths with counts unchanged.
